@Pessimistress 2019-09-26
@jesusbotella Thank you for the detailed probe into this issue.

If I understand it correctly, your proposal is to merge all visible tiles into one dataset (layer) per zoom level.

I have some concerns about introducing the new "combined tile" concept:

- Depending on the size of data, it may actually cost more to create new attributes every time data is updated.
- Not all tile data comes in array format.
- Not all data types can be combined then rendered with a single layer, for example [this](https://github.com/uber/deck.gl/blob/master/examples/website/map-tile/app.js).

I think the root issue in the original approach was that we call `renderSubLayer` even when nothing has changed for a particular tile. The render and subsequent layer diffing both have perf costs. Instead, we could cache and reuse the rendered layer with the tile. I will put together a proof of concept for you to test.