@Pessimistress 2019-09-30
> Could you please elaborate more on the first issue?

If we combine all visible tiles into one layer, then as you pan around and tiles load, that layer needs to recalculate all attributes all the time. The cost of this depends on how much data is visible in one viewport.

> Would you have plans of developing/merging that improvement to the master branch in the near feature?

I'm glad that it works well! I will open a PR. Timing is a little awkward given that we are publishing 7.3.0 today. We can discuss whether this change can be justified as a patch.

> rendering polygons split into different tiles

That is a fair point that this layer does not properly address right now. I think a dedicated `MVTTileLayer` might be the best solution. What are you using for decoding?