@jesusbotella 2019-10-01
> If we combine all visible tiles into one layer, then as you pan around and tiles load, that layer needs to reprocess all geometries that are visible, instead of those of the newly loaded tile. The cost of this depends on how much data is visible in one viewport.

Yes, sure, it makes sense. I understand your point now.

> I'm glad that it works well! I will open a PR. Timing is a little awkward given that we are publishing 7.3.0 today. We can discuss whether this change can be justified as a patch.

Ok, I'll follow that PR to keep in touch with this!

> That is a fair point that this layer does not properly address right now. I think a dedicated MVTTileLayer might be the best solution. What are you using for decoding?

I'm using [`@mapbox/vector-tile`](https://www.npmjs.com/package/@mapbox/vector-tile) to read the protobuf and get a GeoJSON for every feature within the layer.

For the merging approach, I had a features' store and then I was merging polygon vertices grouped by a unique property (defined by the user/developer) shared among features. That was the first approach but we are doing performance reviews and we might change it for a better approach.