@jesusbotella 2019-10-03
Going to close this PR because perfomance improvements have been merged into master. We'll open up a new PR once we have our geometry merge changes ready to be shown and discussed.

Thank you very much!